12 Jan 2020
Added save and load mechanism. Save and load works fine on laptop. Did not check it on phone. Save and load scripts are attached to level manager and saves the state for each star. For now, Save and Load works for only one level and the code design requires us to make seperate script for each level with different save files. 1 save file is of 1KB. Used button for testing that need to be removed.